refactor: test helpers
