+++ Rakefile
@@ -6,7 +6,7 @@ desc 'run tests'
 task default: :test
 
 Rake::TestTask.new do |t|
-  t.libs << 'test'
+  t.libs << ['test', 'test/unit', 'test/integration']
   t.test_files = FileList['test/**/*_spec.rb', 'test/**/test_*.rb']
   t.warning = false
 end
