+++ test/integration/background_color_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'spec_helper'
 
 describe 'user can change background color' do
   describe 'when source is an SVG' do
