+++ test/integration/custom_browserconfig_config_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'spec_helper'
 require 'rexml/document'
 
 describe 'when site overrides source browserconfig' do
