+++ test/integration/empty_site_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'spec_helper'
 
 describe 'empty site' do
   describe 'before processing' do
