+++ test/unit/jekyll/favicon/test_generator.rb
@@ -1,12 +1,20 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'unit_helper'
 
 module Jekyll
   module Favicon
+    # test favicons generator
     class TestGenerator < Minitest::Test
+      include Minitest::Hooks
+
+      def around
+        @site = setup_site
+        super
+        jekyll_execute { Jekyll::Commands::Clean.process @site.config }
+      end
+
       def setup
-        @site = Jekyll::Site.new Jekyll.configuration
         @generator = Generator.new
       end
