+++ test/unit/unit_helper.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+require 'minitest/hooks/test'
+
+# module Context
+#   def around
+#     @site = setup_site
+#     super
+#     jekyll_execute { Jekyll::Commands::Clean.process @site.config }
+#   end
+# end
